[S07P12D105-204] [BE] feat : 로그인 완료 시 반환 정보 추가

See merge request s07-webmobile1-sub2/S07P12D105!54

+++ BE/sellerB/src/main/java/backend/sellerB/dto/ConsultantDto.java
@@ -22,7 +22,7 @@ import java.util.stream.Collectors;
 public class ConsultantDto {
 
 
-    private Integer consultantSeq;
+    private Long consultantSeq;
     private String consultantId;
     private String consultantName;
     private String consultantEmail;
